MAINT: Refactor config facilities
