+++ src/apps/scdetect/CMakeLists.txt
@@ -6,7 +6,9 @@ set(
     amplitudeprocessor.cpp
     app.cpp
     binding.cpp
-    config.cpp
+    config/detector.cpp
+    config/exception.cpp
+    config/template_family.cpp
     datamodel/ddl.cpp
     detail/sqlite.cpp
     detector/arrival.cpp
